+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 
 New features:
 
+ * Add extension for GitHub-style task lists:
+
+   ```
+    * [x] foo
+    * [x] bar
+    * [ ] baz
+   ```
+
  * Renamed structure `MD_RENDERER` to `MD_PARSER` and refactorize its contents
    a little bit. Note this is source-level incompatible and initialization code
    in apps may need to be updated.